+++ src/component/_generated/component.ts
@@ -27,38 +27,49 @@ export type ComponentApi<Name extends string | undefined = string | undefined> =
       add: FunctionReference<
         "mutation",
         "internal",
-        { reactionType: string; targetId: string; userId: string },
+        {
+          namespace?: string;
+          reactionType: string;
+          targetId: string;
+          userId: string;
+        },
         { added: boolean },
         Name
       >;
       getCounts: FunctionReference<
         "query",
         "internal",
-        { targetId: string },
+        { namespace?: string; targetId: string },
         Array<{ count: number; reactionType: string }>,
         Name
       >;
       getUserReactions: FunctionReference<
         "query",
         "internal",
-        { targetId: string; userId: string },
+        { namespace?: string; targetId: string; userId: string },
         Array<string>,
         Name
       >;
       hasUserReacted: FunctionReference<
         "query",
         "internal",
-        { reactionType: string; targetId: string; userId: string },
+        {
+          namespace?: string;
+          reactionType: string;
+          targetId: string;
+          userId: string;
+        },
         boolean,
         Name
       >;
       list: FunctionReference<
         "query",
         "internal",
-        { targetId: string },
+        { namespace?: string; targetId: string },
         Array<{
           _creationTime: number;
           _id: string;
+          namespace?: string;
           reactionType: string;
           targetId: string;
           userId: string;
@@ -68,14 +79,24 @@ export type ComponentApi<Name extends string | undefined = string | undefined> =
       remove: FunctionReference<
         "mutation",
         "internal",
-        { reactionType: string; targetId: string; userId: string },
+        {
+          namespace?: string;
+          reactionType: string;
+          targetId: string;
+          userId: string;
+        },
         { removed: boolean },
         Name
       >;
       toggle: FunctionReference<
         "mutation",
         "internal",
-        { reactionType: string; targetId: string; userId: string },
+        {
+          namespace?: string;
+          reactionType: string;
+          targetId: string;
+          userId: string;
+        },
         { added: boolean },
         Name
       >;